+++ src/main/java/com/elmakers/mine/bukkit/action/builtin/EntityProjectileAction.java
@@ -5,6 +5,7 @@ import com.elmakers.mine.bukkit.api.magic.MageController;
 import com.elmakers.mine.bukkit.api.spell.SpellResult;
 import com.elmakers.mine.bukkit.utility.CompatibilityUtils;
 import com.elmakers.mine.bukkit.utility.ConfigurationUtils;
+import com.elmakers.mine.bukkit.utility.SafetyUtils;
 import org.bukkit.DyeColor;
 import org.bukkit.Location;
 import org.bukkit.configuration.ConfigurationSection;
@@ -200,7 +201,7 @@ public class EntityProjectileAction extends CustomProjectileAction {
             if (velocityOffset != null) {
                 velocity = velocity.add(velocityOffset);
             }
-            entity.setVelocity(velocity);
+            SafetyUtils.setVelocity(entity, velocity);
         }
         Location currentLocation = entity.getLocation();
         if (doTeleport) {