+++ src/main/java/com/elmakers/mine/bukkit/action/builtin/ThrowItemAction.java
@@ -8,6 +8,7 @@ import com.elmakers.mine.bukkit.api.spell.SpellResult;
 import com.elmakers.mine.bukkit.spell.BaseSpell;
 import com.elmakers.mine.bukkit.utility.CompatibilityUtils;
 import com.elmakers.mine.bukkit.utility.NMSUtils;
+import com.elmakers.mine.bukkit.utility.SafetyUtils;
 import org.bukkit.Location;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Item;
@@ -66,7 +67,7 @@ public class ThrowItemAction extends BaseProjectileAction {
         }
         droppedItem.setMetadata("temporary", new FixedMetadataValue(context.getController().getPlugin(), true));
         CompatibilityUtils.ageItem(droppedItem, ageItems);
-        droppedItem.setVelocity(velocity);
+        SafetyUtils.setVelocity(droppedItem, velocity);
 
         track(context, droppedItem);
         return checkTracking(context);