+++ src/main/java/com/elmakers/mine/bukkit/magic/listener/InventoryController.java
@@ -6,6 +6,7 @@ import com.elmakers.mine.bukkit.api.spell.Spell;
 import com.elmakers.mine.bukkit.magic.MagicController;
 import com.elmakers.mine.bukkit.utility.CompleteDragTask;
 import com.elmakers.mine.bukkit.utility.NMSUtils;
+import com.elmakers.mine.bukkit.utility.SafetyUtils;
 import com.elmakers.mine.bukkit.wand.Wand;
 import com.elmakers.mine.bukkit.wand.WandMode;
 import org.bukkit.Location;
@@ -337,7 +338,7 @@ public class InventoryController implements Listener {
                 if (newDrop != null) {
                     Location location = player.getLocation();
                     Item item = location.getWorld().dropItem(location, newDrop);
-                    item.setVelocity(location.getDirection().normalize());
+                    SafetyUtils.setVelocity(item, location.getDirection().normalize());
                 } else {
                     event.setCancelled(true);
                 }