+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/GrenadeSpell.java
@@ -2,6 +2,7 @@ package com.elmakers.mine.bukkit.spell.builtin;
 
 import java.util.Random;
 
+import com.elmakers.mine.bukkit.utility.SafetyUtils;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
 import org.bukkit.configuration.ConfigurationSection;
@@ -49,7 +50,7 @@ public class GrenadeSpell extends BlockSpell
 				return SpellResult.FAIL;
 			}
 			Vector aim = getDirection();
-			grenade.setVelocity(aim);
+			SafetyUtils.setVelocity(grenade, aim);
 			grenade.setYield(size);
 			grenade.setFuseTicks(fuse);
 			grenade.setIsIncendiary(useFire);