+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/LevitateSpell.java
@@ -5,6 +5,7 @@ import com.elmakers.mine.bukkit.api.magic.MageController;
 import com.elmakers.mine.bukkit.api.spell.Spell;
 import com.elmakers.mine.bukkit.effect.builtin.EffectRing;
 import com.elmakers.mine.bukkit.magic.MagicController;
+import com.elmakers.mine.bukkit.utility.SafetyUtils;
 import com.elmakers.mine.bukkit.wand.Wand;
 import com.elmakers.mine.bukkit.block.MaterialAndData;
 import com.elmakers.mine.bukkit.utility.CompatibilityUtils;
@@ -355,7 +356,7 @@ public class LevitateSpell extends TargetingSpell implements Listener
         if (mountEntity != null) {
             ArmorStand activeArmorStand = armorStand != null ? armorStand : (mountEntity instanceof ArmorStand ? (ArmorStand)mountEntity : null);
             if (activeArmorStand != null) {
-                activeArmorStand.setVelocity(direction);
+                SafetyUtils.setVelocity(activeArmorStand, direction);
                 CompatibilityUtils.setYawPitch(activeArmorStand, location.getYaw() + (float)yawAmount, location.getPitch());
                 if (pitchAmount != 0) {
                     if (armorStandHead == null && useHelmet) {
@@ -368,10 +369,10 @@ public class LevitateSpell extends TargetingSpell implements Listener
                     }
                 }
             } else {
-                mountEntity.setVelocity(direction);
+                SafetyUtils.setVelocity(mountEntity, direction);
             }
         } else {
-            player.setVelocity(direction);
+            SafetyUtils.setVelocity(player, direction);
         }
 
         if (effectParticle != null) {
@@ -923,12 +924,12 @@ public class LevitateSpell extends TargetingSpell implements Listener
 
             if (mountEntity != null) {
                 if (armorStand != null) {
-                    armorStand.setVelocity(velocity);
+                    SafetyUtils.setVelocity(armorStand, velocity);
                 } else {
-                    mountEntity.setVelocity(velocity);
+                    SafetyUtils.setVelocity(mountEntity, velocity);
                 }
             } else {
-                player.setVelocity(velocity);
+                SafetyUtils.setVelocity(player, velocity);
             }
         }
         if (flight) {