+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/TossSpell.java
@@ -1,5 +1,6 @@
 package com.elmakers.mine.bukkit.spell.builtin;
 
+import com.elmakers.mine.bukkit.utility.SafetyUtils;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.configuration.ConfigurationSection;
@@ -57,7 +58,7 @@ public class TossSpell extends BrushSpell
 			}
 			registerForUndo(block);
 			block.setDropItem(false);
-			block.setVelocity(direction);	
+			SafetyUtils.setVelocity(block, direction);
 		}
 
 		registerForUndo();